// ui/src/generated/PackageEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PackageSource } from "./PackageSource";

export type PackageEntry = { name: string, version: string, title: string | null, built: string | null, library_path: string | null, library_index: number | null, first_in_library_path: boolean, loaded_version: string | null, loaded_path: string | null, loaded_from_library: boolean, attached: boolean, source: PackageSource | null, };